sfinit, bfinit, blkfc1, blkct, inpnv, symfct implemented
fix bug 6401 :
test_run("sparse","bug_6401",["no_check_error_output" ]);

+++ scilab/modules/sparse/sci_gateway/cpp/sci_lufact.cpp
@@ -105,7 +105,7 @@ types::Function::ReturnValue sci_lufact(types::typed_list &in, int _iRetCount, t
         return types::Function::Error;
     }
 
-    nonZeros = pSpIn->nonZeros();
+    nonZeros = (int)pSpIn->nonZeros();
     dbl = new double[nonZeros];
     colPos = new int[nonZeros];
     itemsRow = new int[m];
